first version (unmodified paged geom source, grass is on road, wrong cache dir, maybe other problems)

+++ source/paged-geom/RandomTable.h
@@ -91,7 +91,7 @@ inline void RandomTable::generateRandomNumbers()
 	MTRand mtrand(time(NULL));
 #endif
 	for(unsigned long i = 0; i < tableSize; i++)
-		table[i] = mtrand.rand();
+		table[i] = (float)mtrand.rand();
 #endif //USE_OGRE_RANDOM
 }
